% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryInput.R
\name{choose.dir.linux}
\alias{choose.dir.linux}
\alias{choose.dir.windows}
\title{The linux version of the choose folder}
\usage{
choose.dir.linux(default = NA, caption = NA)

choose.dir.windows(default = NA, caption = NA, useNew = TRUE)
}
\value{
A length one character vector, character NA if 'Cancel' was selected.

A length one character vector, character NA if 'Cancel' was selected.
}
\description{
The linux version of the choose folder

The windows version of the choose folder
}
\seealso{
\code{\link{choose.dir}}

\code{\link{choose.dir}}
}
