% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryInput.R
\name{choose.dir}
\alias{choose.dir}
\title{Choose a Folder Interactively

Display an OS-native folder selection dialog under Mac OS X, Linux GTK+ or
Windows.}
\usage{
choose.dir(default = NA, caption = NA, useNew = TRUE)
}
\arguments{
\item{default}{which folder to show initially}

\item{caption}{the caption on the selection dialog}

\item{useNew}{boolean, selects the type of dialog shown in windows}
}
\value{
A length one character vector, character NA if 'Cancel' was selected.
}
\description{
Choose a Folder Interactively

Display an OS-native folder selection dialog under Mac OS X, Linux GTK+ or
Windows.
}
\details{
Uses an Apple Script, Zenity or Windows Batch script to display an OS-native
folder selection dialog.

For Apple Script, with \code{default = NA}, the initial folder selection
is determined by default behavior of the "choose folder" script. Otherwise,
paths are expanded with \code{\link{path.expand}}.

For Linux, with \code{default = NA}, the initial folder selection is
determined by defaul behavior of the zenity script.

The new windows batch script allows both initial folder and caption to be set.
In the old batch script for Windows the initial folder is always ignored.
}
