% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryInput.R
\name{choose.dir.windows}
\alias{choose.dir.windows}
\title{The windows version of the choose folder}
\usage{
choose.dir.windows(default = NA, caption = NA, useNew = TRUE)
}
\value{
A length one character vector, character NA if 'Cancel' was selected.
}
\description{
The windows version of the choose folder
}
\seealso{
\code{\link{choose.dir}}
}
