% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runDirinputExample}
\alias{runDirinputExample}
\title{Runs a demo app with the code

See the code in the example_application folder}
\usage{
runDirinputExample()
}
\description{
Runs a demo app with the code

See the code in the example_application folder
}
