% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runDirinputExample}
\alias{runDirinputExample}
\title{Runs a demo app with the code}
\usage{
runDirinputExample()
}
\description{
Runs a demo app with the code
}
\details{
Runs a demo app with the code. See the code in the example_application folder for
how this is accomplished.
}
